ajout des cases vide dans le planning

+++ php/profil.php
@@ -97,13 +97,17 @@ if ($_SESSION['login'] == false) {
                     }
                 }
             }
-            if (isset($_POST['delete_user'])) {
-                if ($_SESSION['login']) {
-                    if ($_POST['delete_user']) {
-                        $delete_user = $bdd->prepare("DELETE FROM utilisateurs WHERE id = ?");
-                        $delete_user->execute([$_SESSION['id']]);
-                        session_destroy();
-                        header('Location: planning.php');
+            if ($_SESSION['login'] == 'admin') {
+                echo "<p><i class='fa-solid fa-triangle-exclamation'></i>&nbspIMPOSSIBLE de supprimer ce compte.</p>";
+            } else {
+                if (isset($_POST['delete_user'])) {
+                    if ($_SESSION['login']) {
+                        if ($_POST['delete_user']) {
+                            $delete_user = $bdd->prepare("DELETE FROM utilisateurs WHERE id = ?");
+                            $delete_user->execute([$_SESSION['id']]);
+                            session_destroy();
+                            header('Location: planning.php');
+                        }
                     }
                 }
             }
